style(difficulty): move updater in its own class
